moved sdmo aggregator to a new file
