+++ src/mobgap/aggregation/__init__.py
@@ -1,6 +1,7 @@
 """Algorithms to aggregate results across walking bouts."""
 
-from mobgap.aggregation._mobilised_aggregator import MobilisedAggregator, SDMOAggregator
+from mobgap.aggregation._mobilised_aggregator import MobilisedAggregator
+from mobgap.aggregation._signal_based_aggregator import SDMOAggregator
 from mobgap.aggregation._threshold_check import apply_thresholds, get_mobilised_dmo_thresholds
 
 __all__ = ["MobilisedAggregator", "SDMOAggregator", "apply_thresholds", "get_mobilised_dmo_thresholds"]
